fix(test): apply 3 review findings from 1b1 walkthrough
- Extract make_message() to tests/core/conftest.py (shared helper)
- Update ADR-001 frontmatter description to mention scope_id amendment
- Add exact pool_id assertions in TestScopeIsolatedPools
