+++ docs/architecture/adr/001-routing-key-platform-bot-id-user-id.mdx
@@ -1,6 +1,6 @@
 ---
 title: "ADR-001: RoutingKey(platform, bot_id, user_id) replaces BindingKey(channel, user_id)"
-description: Adopt a three-dimensional typed routing key to support multi-bot, multi-channel dispatch with full attribution.
+description: "Adopt a three-dimensional typed routing key to support multi-bot, multi-channel dispatch. Amended by #125: third field renamed from user_id to scope_id for conversation-scope routing."
 ---
 
 ## Status
